Move examination classes to separate package
