+++ src/main/java/de/wevsvirushackathon/coronareport/examination/RespiratoryInfectionQuery.java
@@ -1,4 +1,4 @@
-package de.wevsvirushackathon.coronareport.diary;
+package de.wevsvirushackathon.coronareport.examination;
 
 import de.wevsvirushackathon.coronareport.client.Client;
 import lombok.Data;
